add authorization check

+++ searchManager/WebContent/js/bigbets/relevancy.js
@@ -21,6 +21,16 @@
 	var bqFacetValuesPageSize = 5;
 	var bqSearchText = "Enter Field Value";
 
+	var allowModify = true;
+	
+	var getPermission = function() {
+		UtilityServiceJS.hasPermission('CREATE_RULE', {
+		callback:function(data){
+			allowModify = data;
+		}
+	});
+	};
+
 	/** BELOW: BF */
 	var setupFieldS4 = function(field){
 		$('div[id="' + field.id + '"] a.editIcon, div[id="' + field.id + '"] input[type="text"]').qtip({
@@ -618,11 +628,11 @@
 		$('div#relevancy .saveIcon').off().on({
 			click:function(e){
 				var field = getRelevancyField(e);
-				if (!e.data.locked) 
+				if (!e.data.locked || allowModify ) 
 					addRuleFieldValue(field.id, field.value);
 			},
 			mouseenter: showHoverInfo
-		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default")});
+		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default") || !allowModify});
 
 		showGraph = function(e){
 			var field = getRelevancyField(e);
@@ -798,6 +808,7 @@
 	};
 
 	var cloneRule = function(e){
+		if (e.data.locked || !allowModify) return;
 		$(this).qtip({
 			content: {
 				text: $('<div/>'),
@@ -897,7 +908,7 @@
 	};
 
 	var updateRule = function(e){
-		if (e.data.locked) return;
+		if (e.data.locked || !allowModify) return;
 
 		var unSaved = getUnSavedRelevancyFields();
 		var ruleName = $.trim($('div#relevancy input[id="name"]').val()); 
@@ -968,7 +979,7 @@
 	};
 
 	var deleteRule = function(e) { 
-		if (!e.data.locked && confirm(deleteRuleConfirmText)){
+		if (!e.data.locked && allowModify && confirm(deleteRuleConfirmText)){
 			RelevancyServiceJS.deleteRule(selectedRule.ruleId,{
 				callback: function(code){
 					if (code > 0) {
@@ -1021,7 +1032,7 @@
 			showOn: "both",
 			buttonImage: "../images/icon_calendar.png",
 			buttonImageOnly: true,
-			disabled: selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default"),
+			disabled: selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default") || !allowModify,
 			onSelect: function(selectedDate) {
 				var option = this.id == "startDate" ? "minDate" : "maxDate",
 						instance = $(this).data("datepicker"),
@@ -1039,16 +1050,17 @@
 		$("#saveBtn").off().on({
 			click: updateRule,
 			mouseenter: showHoverInfo
-		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default")});
+		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default") || !allowModify});
 
 		$("#cloneBtn").off().on({
-			click: cloneRule
-		},{locked:false});
+			click: cloneRule,
+			mouseenter: showHoverInfo
+		},{locked:!allowModify});
 
 		$("#deleteBtn").off().on({
 			click: deleteRule,
 			mouseenter: showHoverInfo
-		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default")});
+		},{locked:selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default") || !allowModify});
 		
 		$("a#downloadIcon").download({
 			headerText:"Download Ranking Rule",
@@ -1094,7 +1106,7 @@
 
 		$('#auditIcon').on({
 			click: showAuditList
-		}, {locked: false, type:moduleName, ruleRefId: selectedRule.ruleId, name: selectedRule.ruleName});
+		}, {locked: !allowModify, type:moduleName, ruleRefId: selectedRule.ruleId, name: selectedRule.ruleName});
 
 	};
 
@@ -1389,7 +1401,7 @@
 			pageSize: keywordInRulePageSize,
 			headerText : "Using This Rule",
 			searchText : "Enter Keyword",
-			showAddButton: !selectedRuleStatus.locked && !$.endsWith(selectedRule.ruleId, "_default"),
+			showAddButton: !selectedRuleStatus.locked && !$.endsWith(selectedRule.ruleId , "_default") || allowModify,
 			itemDataCallback: function(base, keyword, page){
 				RelevancyServiceJS.getAllKeywordInRule(selectedRule.ruleId, keyword, page, keywordInRulePageSize, {
 					callback: function(data){
@@ -1408,7 +1420,7 @@
 
 				base.$el.find('#itemPattern' + suffixId + ' div.itemLink a#delete' + suffixId).on({
 					click: function(e){
-						if (!e.data.locked && confirm('Delete "' + name + '" in ' + selectedRule.ruleName  + '?'))
+						if (!e.data.locked && allowModify && confirm('Delete "' + name + '" in ' + selectedRule.ruleName  + '?'))
 							RelevancyServiceJS.deleteKeywordInRule(selectedRule.ruleId, name,{
 								callback:function(code){
 									showActionResponse(code, "delete", name);
@@ -1420,7 +1432,7 @@
 							});
 					},
 					mouseenter: showHoverInfo
-				},{locked: selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default")});
+				},{locked: selectedRuleStatus.locked || $.endsWith(selectedRule.ruleId, "_default") || !allowModify});
 			},
 			itemAddCallback: function(base, keyword){
 				RelevancyServiceJS.addKeywordToRule(selectedRule.ruleId, keyword, {
@@ -1437,6 +1449,7 @@
 	};
 
 	$(document).ready(function() { 
+		getPermission();
 		showRelevancy();
 		getRelevancyRuleList();
 		getRelevancyRuleKeywordList();
